+++ workers/unity/Packages/io.improbable.gdk.mobile/Editor/DeviceLaunchConfig.cs
@@ -1,5 +1,4 @@
 using System;
-using UnityEngine;
 
 namespace Improbable.Gdk.Mobile
 {
@@ -13,37 +12,20 @@ namespace Improbable.Gdk.Mobile
 
     public struct DeviceLaunchConfig
     {
-        public readonly string deviceName;
-        public readonly string deviceId;
-        public readonly string prettyDeviceType;
-        public readonly DeviceType deviceType;
+        public readonly string DeviceName;
+        public readonly string DeviceId;
+        public readonly string PrettyDeviceType;
+        public readonly DeviceType DeviceType;
 
         private readonly Action<DeviceLaunchConfig, MobileLaunchConfig> launchAction;
 
         public DeviceLaunchConfig(string deviceName, string deviceId, DeviceType deviceType, Action<DeviceLaunchConfig, MobileLaunchConfig> launchAction)
         {
-            this.deviceName = deviceName;
-            this.deviceId = deviceId;
-            this.deviceType = deviceType;
+            this.DeviceName = deviceName;
+            this.DeviceId = deviceId;
+            this.DeviceType = deviceType;
             this.launchAction = launchAction;
-            prettyDeviceType = GetPrettyDeviceType(deviceType);
-        }
-
-        public static string GetPrettyDeviceType(DeviceType deviceType)
-        {
-            switch (deviceType)
-            {
-                case DeviceType.AndroidDevice:
-                    return "Android Device";
-                case DeviceType.AndroidEmulator:
-                    return "Android Emulator";
-                case DeviceType.iOSDevice:
-                    return "iOS Device";
-                case DeviceType.iOSSimulator:
-                    return "iOS Simulator";
-                default:
-                    throw new ArgumentException($"Invalid Device Type: {deviceType}");
-            }
+            PrettyDeviceType = deviceType.ToPrettyDeviceType();
         }
 
         public void Launch(MobileLaunchConfig mobileLaunchConfig) => launchAction(this, mobileLaunchConfig);